Changed LWJGL to version 2 instead of 3, apparently their new release isn't full implemented... Started working on the driver.

+++ GPGPU_Comparisons/src/Vec_Add.cl
@@ -1,4 +1,4 @@
-kernel void Vec_Add(global const int* a, global const int* b, global int* answer, global int n) {
+kernel void Vec_Add(global const double* a, global const double* b, global int* answer, global int n) {
   unsigned int xid = get_global_id(0);
   if(xid < n){
 	  answer[xid] = a[xid] + b[xid];
